data_collection/requirements.txt
python-dotenv
requests
networkx
